backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic[email]>=2.7.0
pydantic-settings>=2.3.0
supabase>=2.5.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
httpx>=0.27.0
APScheduler>=3.10.0
sendgrid>=6.11.0
email-validator>=2.1.0
python-multipart>=0.0.9
weasyprint>=61.0
